move hid_data to platform/hid-data.h
