+++ src/linux/backend-v4l2.cpp
@@ -3,6 +3,7 @@
 
 #include "backend-v4l2.h"
 #include <src/platform/command-transfer.h>
+#include <src/platform/hid-data.h>
 #include <src/core/time-service.h>
 #include <src/core/notification.h>
 #include "backend-hid.h"
