+++ src/types.h
@@ -59,34 +59,8 @@ using std::abs;
 #endif
 
 
-namespace librealsense
-{
-#pragma pack (push, 1)
+namespace librealsense {
 
-    struct hid_data
-    {
-        short x;
-        byte reserved1[2];
-        short y;
-        byte reserved2[2];
-        short z;
-        byte reserved3[2];
-    };
-
-    // D457 dev
-    struct hid_mipi_data
-    {
-        uint8_t typeID;
-        uint8_t skip1;
-        uint64_t hwTs;
-        int16_t x;
-        int16_t y;
-        int16_t z;
-        uint64_t hwTs2;
-        uint64_t skip2;
-    };
-
-#pragma pack(pop)
 
     static const double TIMESTAMP_USEC_TO_MSEC = 0.001;
